+++ crates/utils/sov-evm-test-utils/contracts/artifacts/ERC20.bin
