+++ crates/utils/sov-evm-test-utils/contracts/artifacts/IERC20.abi
