+++ crates/utils/sov-evm-test-utils/contracts/artifacts/SimpleStorage.abi
