+++ crates/utils/sov-evm-test-utils/contracts/artifacts/SimpleStorage.bin
