+++ crates/utils/sov-evm-test-utils/contracts/build.sh
