+++ scripts/switcheroo/presets/default.json
@@ -25,6 +25,7 @@
     "crates/module-system/sov-modules-api",
     "crates/module-system/sov-address",
     "crates/utils/sov-test-utils",
+    "crates/utils/sov-evm-test-utils",
     "crates/module-system/module-implementations/sov-accounts",
     "crates/module-system/module-implementations/sov-bank",
     "crates/module-system/module-implementations/sov-chain-state",
