+++ scripts/switcheroo/presets/ethereum.json
@@ -2,5 +2,6 @@
 	"crates/full-node/sov-ethereum",
 	"crates/full-node/sov-eth-dev-signer",
 	"crates/utils/sov-evm-soak-testing",
+	"crates/utils/sov-evm-test-utils",
 	"crates/module-system/module-implementations/sov-evm"
 ]
